restructure components and styling
