+++ src/components/HealthInformation/HealthInformation.scss
@@ -1,47 +1,46 @@
-@import '../variables';
+@import "../../styles/variables";
 
 @media only screen and (min-width: 768px) and (max-width: 1024px) {
-  .health-container{
+  .health-container {
     display: flex;
     flex-direction: column;
     align-items: center;
-    form{
-      button{
-        border:none;
+    form {
+      button {
+        border: none;
         background-color: $blue;
-        color:$white;
-        padding:5px;
+        color: $white;
+        padding: 5px;
         font-size: 1em;
       }
       display: flex;
       flex-direction: column;
       align-items: center;
-      div{
+      div {
         width: 150%;
         display: flex;
-        margin:10px 0;
-        label{
+        margin: 10px 0;
+        label {
           width: 20%;
           text-align: right;
         }
-        input{
-          margin:0 10px;
+        input {
+          margin: 0 10px;
           width: 100%;
           font-size: 1.2em;
-          padding:10px;
+          padding: 10px;
         }
-        
-        textarea{
-          margin:0 10px;
+
+        textarea {
+          margin: 0 10px;
           width: 100%;
           font-size: 1.2em;
-          padding:10px;
+          padding: 10px;
         }
-        select{
-          margin:0 10px;
+        select {
+          margin: 0 10px;
           width: 100%;
         }
-        
       }
     }
   }
