+++ src/containers/Appointments/Appointments.scss
@@ -1,4 +1,4 @@
-@import "../variables";
+@import "../../styles/variables";
 
 @media only screen and (min-width: 768px) and (max-width: 1024px) {
   .container-appointments {
