+++ src/containers/BookAppointment/BookAppointment.js
@@ -1,11 +1,11 @@
 import React, { useState } from 'react';
 import DateTimePicker from 'react-datetime-picker';
 import 'react-calendar/dist/Calendar.css';
-import '../styles/tablets/book.scss';
+import './BookAppointment.scss';
 import { connect } from 'react-redux';
 import { useParams, useLocation } from 'react-router-dom';
 import PropTypes from 'prop-types';
-import { bookAppointment } from '../store/actions/appointments';
+import { bookAppointment } from '../../store/actions/appointments';
 
 const BookAppointment = ({ bookAppointment, userid }) => {
   const [info, setInfo] = useState('');
