+++ src/containers/Home/Home.scss
@@ -1,8 +1,7 @@
-@import '../variables';
+@import '../../styles/variables';
 
 @media only screen and (min-width: 768px) and (max-width: 1024px) {
-  .welcome{
-  }
+
   .menu{
     border-bottom: 1px solid #cccccc;
     li{
