+++ src/containers/Specialities/Specialities.js
@@ -1,10 +1,10 @@
 import React, { useEffect } from 'react';
 import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
-import { getSpecialities } from '../store/actions/specialities';
-import Loading from '../components/Loading';
-import '../styles/tablets/specialities.scss';
-import Speciality from '../components/Speciality';
+import { getSpecialities } from '../../store/actions/specialities';
+import Loading from '../../components/Loading/Loading';
+import './Specialities.scss';
+import Speciality from '../../components/Speciality/Speciality';
 
 const Specialities = ({ specialities, getSpecialities, isLoading }) => {
   useEffect(() => {
