+++ src/index.js
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom';
 import './index.css';
 import { BrowserRouter } from 'react-router-dom';
 import { Provider } from 'react-redux';
-import App from './containers/App';
+import App from './containers/App/App';
 import * as serviceWorker from './serviceWorker';
 import store from './store';
